refactor(calibration): split MasterDetail into useMasterDetail hook, master-detail-model.ts builders, and MasterArchiveFlow (CU-22)
588L → ~200L render; fmtBytes → formatBytes canon from @/lib/format; archive modal pair extracted to MasterArchiveFlow.
